Delete test scenario
----
# New Feature

- REST API Delete scenario (soft delete)

---
# Updated

- Documentation
  - Routing
  - WBS
- Unit tests
- Scenario model
- Move CampaignScenario to BaseScenario
- Move FillCampaignResult to CreateUpdateModel

# Fixed

- Unit test (before all and teardown)
