+++ app/schema/campaign/__init__.py
@@ -1,0 +1,2 @@
+# -*- Product under GNU GPL v3 -*-
+# -*- Author: E.Aivayan -*-
