+++ app/schema/ticket_schema.py
@@ -9,6 +9,14 @@ from app.schema.status_enum import TicketType
 
 
 class Ticket(BaseModel):
+    """
+    Attributes:
+        status
+        reference
+        description
+        created
+        updated
+    """
     status: str
     reference: str
     description: str
